requirements.txt
django==1.11.6
djangorestframework==3.7.1
pillow